pitch project functionality added
