+++ backend/app1/admin.py
@@ -8,6 +8,8 @@ admin.site.register(PassionUser)
 admin.site.register(SuperUser)
 admin.site.register(Client)
 admin.site.register(Creator)
+admin.site.register(UserSampleWorkTable)
 admin.site.register(Project)
 admin.site.register(Role)
 admin.site.register(Application)
+admin.site.register(ProjectSampleWorkTable)
