+++ backend/backend/settings.py
@@ -99,6 +99,11 @@ DATABASES = {
 
 AUTH_USER_MODEL = "app1.BaseUser"
 
+REST_FRAMEWORK = {
+   'DEFAULT_PAGINATION_CLASS': 'rest_framework.pagination.PageNumberPagination',
+   'PAGE_SIZE': 10
+}
+
 # Password validation
 # https://docs.djangoproject.com/en/5.0/ref/settings/#auth-password-validators
